the prolog dialog must be done only once
Let's revert previous change partly: one http client is created in the
agent for the prolog, and the needed parameters are passed to each
task so as let them create another one if needed.

Additionaly, the server response to the prolog is passed to task
isEnabled() method, allowing earlier checks about the usefulness
of running it.