don't reencode if perl flagged string as utf8
If Perl flagged a string as UTF8 internally, there is no reason
to reencode it. A best, this will hide a bug somewhere else.

In my opinion, we should no try to encode string at all.

To know if a string is flagged as UTF8, one can use Devel::Peek.