Win32: properly support non ASCII characters
We convert the file name from UTF-8 to the local encoding. This
works fine if all the caracter are in the local encoding.

To support all the cases, we should use the Win32 UTF16 API, but
this requires to heavily patch File::Copy::Recursive and
probably some 3rd party lib like File::Copy.

See: Win32::Unicode::File and Win32API::File