unicode fix around Win32::OLE
We configured Win32::OLE to use UTF8 with the CP => CP_UTF8 parameter.
The string are stored as Perl internal string but without the utf8 flag.

Since we know it's UTF8, we call utf8::upgrade to add the internal UTF8
flag on them.

Win32::OLE may returns non "string" item like array so we need to be
cleaver a bit.

This changes also add a fix for the case Win32::TieRegistry returns string with the UTF8 flag. In this case we won't reencode them.