fix: File::Path::mkpath error handling
RHEL5's File::Path::mkpath() doesn't support the error parameter.
According to the documentation, if we don't pass this parameter,
File::Path::mkpath() will die() in case of error. This is by
the way what we want/do.

So, I drop this {error => \my $err}, this will avoid a error message
on RHEL5 and will simplify the current code (error → die).

closes: #1817

Reported-by: Jason Smith <smithj4@bnl.gov>