Work around Net::IP bug 71510
If the server send a bad IP range with 0.0.0.0, Net::IP->new() do not
generate a valid Net::IP object. We get an exception if we try to use
it:

$perl -MNet::IP -e '$ip = Net::IP->new("0.0.0.0-0.0.0.0"); print $ip->size()'
Can't call method "is_zero" on an undefined value at /usr/share/perl5/Math/BigInt.pm line 1197.
$perl -MNet::IP -e '$ip = Net::IP->new("1.0.0.0-1.0.0.0"); print $ip->size()'

The bug is already open on Net::IP:

https://rt.cpan.org/Public/Bug/Display.html?id=71510

closes: #1880
see: #1881

Reported-by: Walid Nouh <walid@wawax.info>