drop vlan-specific SNMP sessions
This method is only used to retrieve mac addresses associated with a
port, without any hint about wether they correspond to devices directly
connected to the port itself, or indirectly connected through another
networking device. If those mac addresses are only visible in a
specific VLAN, and not from default one, they are likely to be
associated for routing only.

Moreover, this method missed information available through the initial
session, and was relying on community, meaning it was specific to SNMP
v1 or v2.