Replace use of "install -d"
The ```-d (--directory)``` option to ```install``` is not available on all platforms (in particular older Unixes). It behaves the same as ```mkdir -p```, except that permissions are unpredicatable due to potential interference from users' umask.

The usage in Makefile.PL here is always ```install -d -m 755``` so I'm proposing to replace these calls by ```mkdir -p``` and ```chmod 755```.