feat: limit next run delay reduc
Limit next run random delay reduction to max 1/6 of the delay if less then 6 hours,
limit to max an hour for delay from 6 hours to 24 hours,
and limit to max 1/24 of the delay for delay greater than a day.

This would keep enough delay randomization to avoid mass agent server connection and
keep next seen run more coherent with the requested delay.